Remove server files
