--- package.json
@@ -1,34 +1,0 @@
-{
-  "name": "node-sequelize",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "chalk": "^4.1.1",
-    "connect-livereload": "^0.6.1",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "livereload": "^0.9.3",
-    "mysql": "^2.18.1",
-    "mysql2": "^2.2.5",
-    "node-fetch": "^3.2.0",
-    "nodemon": "^2.0.7"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.17",
-    "@babel/register": "^7.12.13",
-    "eslint": "^7.20.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-cypress": "^2.11.1",
-    "eslint-plugin-import": "^2.22.1"
-  }
-}
