--- server/queries/index.js
@@ -1,7 +1,0 @@
-import diningHallQuery from './diningHallQuery.js';
-import mealsQuery from './mealsQuery.js';
-
-export default {
-  meals: mealsQuery,
-  halls: diningHallQuery
-};
