--- server/queries/mealsQuery.js
@@ -1,1 +1,0 @@
-export default `SELECT * FROM meals WHERE meal_category = ?;`;
